Update Readme Fixes #18

+++ AlinasMapMod/Settings.cs
@@ -2,6 +2,6 @@ namespace AlinasMapMod {
   class Settings {
     public bool EnableDeliveries {get; set;} = true;
     public bool FreeMilestones {get; set;} = false;
-    public float DeliveryCarCountMultiplier { get; set; } = 1;
+    public int DeliveryCarCountMultiplier { get; set; } = 1;
   }
 }
